+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 3. 将金仕达黄金接口剥离到vnpy_ksgold项目中
 4. 将投资组合策略模块剥离到vnpy_portfoliostrategy项目中
 5. 将Excel RTD模块剥离到vnpy_excelrtd项目中
-
+6. 将本地仿真模拟交易模块剥离到vnpy_paperaccount项目中
 
 # 2.7.0版本