+++ CHANGELOG.md
@@ -6,10 +6,12 @@
 
 # 修复
 1. 修复部分C++接口模块剥离后，遗留的安装脚本编译代码导致的报错问题
+2. 修复vnpy_xtp订阅深交所行情后，可能出现的闪退问题
 
 # 调整
 1. 移除老版本基于requests库的RestClient客户端
 2. 移除老版本基于websocket-client库的WebsocketClient客户端
+3. vnpy_tts增加对上交所和深交所股票模拟交易的支持
 
 # 剥离
 1. 将华鑫奇点交易接口剥离到vnpy_tora项目中，并升级到4.0版本